chore : add pessimistic write lock on event ticketprice to lock the row from updates chore : add reserve method in booking service

+++ src/main/java/com/tiko/tiko/Events/DTO/EventDetailsDTO.java
@@ -9,7 +9,7 @@ public record EventDetailsDTO(
         String name,
         LocalDateTime date,
         String location,
-        Long capacity,
+        int capacity,
         String description,
         EventStatus status,
         String imageUrl,
